requirements.txt
Flask==2.3.2
pandas==2.0.1
plotly==5.10.0
google-auth==2.17.1
google-analytics-data==0.9.1